Move main into a cmd/operator package
