+++ cmd/operator/display/root.go
