+++ cmd/operator/register/root.go
